Force the DOM to always rerender
This fixes a bug that was causing the DOM to not rerender. So if a build
status changed it wouldn't be reflected because none of the data had
changed.

I think a better solution to this would be to see if there is some other
field that changes that we can use from github. Perhaps a workflow run
datetime or ideally a field that says if the job passed/failed.