Add the build status as a query parameter to the badge image
It seems that the DOM not rendering and the browser caching the image
were causing it to not refresh. I believe this fix causes two things:

1. When the setter is called the virtual DOM is updated but only the actual
DOM is update if there was a change. This query param will change when
the build was a 'success' or 'failure'. So then the DOM will update in
step.

2. The browser will cache the image (5 minutes is the Cache-Control
header git sends with the image). So it will at least be this long
before it refreshes. Instead, to cause it to not cache we add the build
status. This makes the browser treat it as two separate URLs so it will
end up caching both and reference the correct one when requested.

- Rename "badge info" code references to "build info".