Don't display mint dashboard card
I don't think this dashboard is an appropriate place for this card. It
was a good experiment to play around with runtime secrets, but the point
of the dashboard is for public information.
Additionally, the backend job for this isn't even working right now due
to a bug or two in the underlying mint lib.