Use node 16.13.1 in build image
There was an error in the build when running `npm ci`. Not exactly sure
what was causing it, but it seemed to be due to the fact that we were on
an old version of node. We've been getting warnings for a while from npm
that we're on an unsupported Node.js version.

This also moves from using the root user to an unprivileged user in the
build container which is a best practice anyways.