Support CI builds in PRs
We need to use the branch that the PR is using. We pull this from the
github environment variable and then clone the repo using that branch.

We also don't want to deploy so we need some way to signal that from the
PR build. We do this by specifying a build task and don't deploy when we
only want to run tests.