Update Dockerfile for Ubuntu 24 migration
After migrating to Ubuntu 24 there were a few things we needed to
change. Mainly the apt dependencies needed to be upgraded since the
versions were no longer correct and some packages migrated names.
Notably, awscli is no longer available as an apt package so we needed to
install it manaually. The way we created a user also needed to be
updated.